Task about calendar start day.
The task given as a quiz during the class seemed to me a little bit challenging for first year beginner students, even for myself. So, I decided to hand code myself, and here I add it for future reference. This task is quite interesting and somehow challenging. There are variety of ways to accomplish it.